requirements.txt
aiosqlite==0.19.0
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.3.1
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.2.2
fastapi==0.103.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
kombu==5.4.2
numpy==2.2.3
pandas==2.2.3
Pillow==10.0.0
prompt_toolkit==3.0.50
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.1
redis==4.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.12.2
tzdata==2025.1
uuid==1.30
uvicorn==0.23.2
vine==5.1.0
wcwidth==0.2.13